move BenchmarkDecode into integration_test
